# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/driver/profile": {
      "post": {
        "operationId": "VerificationController_upsertProfile",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initialize or update captain profile",
        "tags": [
          "Driver Onboarding & Verification"
        ]
      }
    },
    "/v1/driver/vehicles": {
      "post": {
        "operationId": "VerificationController_registerVehicle",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a vehicle under driver profile",
        "tags": [
          "Driver Onboarding & Verification"
        ]
      }
    },
    "/v1/driver/documents/upload-url": {
      "post": {
        "operationId": "VerificationController_getUploadUrl",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get pre-signed upload URL for private verification document",
        "tags": [
          "Driver Onboarding & Verification"
        ]
      }
    },
    "/v1/driver/documents": {
      "post": {
        "operationId": "VerificationController_submitDocument",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit KYC / vehicle document for verification review",
        "tags": [
          "Driver Onboarding & Verification"
        ]
      }
    },
    "/v1/driver/verification": {
      "get": {
        "operationId": "VerificationController_getVerificationStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current captain verification tier and missing documents",
        "tags": [
          "Driver Onboarding & Verification"
        ]
      }
    },
    "/v1/admin/verifications": {
      "get": {
        "operationId": "VerificationController_listVerifications",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List pending verification documents",
        "tags": [
          "Driver Onboarding & Verification"
        ]
      }
    },
    "/v1/admin/verifications/{id}/approve": {
      "post": {
        "operationId": "VerificationController_approveVerification",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve driver verification document",
        "tags": [
          "Driver Onboarding & Verification"
        ]
      }
    },
    "/v1/admin/verifications/{id}/reject": {
      "post": {
        "operationId": "VerificationController_rejectVerification",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject driver verification document with reason",
        "tags": [
          "Driver Onboarding & Verification"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wasel API",
    "description": "Wasel Logistics & Delivery Platform Modular Monolith API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "idempotency-key": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key"
      }
    },
...
  }
}